filtering list and showing correct followers/following count
