+++ src/components/ConnectionsListProfiles.tsx
@@ -18,16 +18,12 @@ enum ListStatus {
 
 interface ConnectionsListProfilesProps {
   userId: string;
-  listStatus: ListStatus;
-  followedByDisplayUser: Record<string, RelationshipState>;
-  followingDisplayUser: Record<string, RelationshipState>;
+  connectionsList: User[];
 }
 
 const ConnectionsListProfiles = ({
   userId,
-  listStatus,
-  followedByDisplayUser,
-  followingDisplayUser,
+  connectionsList,
 }: ConnectionsListProfilesProps): JSX.Element => {
   const users: Record<string, User> = useAppSelector(
     (state) => state.profiles?.profiles || {}
@@ -36,29 +32,6 @@ const ConnectionsListProfiles = ({
   const followedByCurrentUser = useAppSelector(
     (state) => (userId && state.graphs.following[userId]) || {}
   );
-
-  const profileForId = (userId: string): User =>
-    users[userId] || {
-      contentHash: "",
-      url: "",
-      announcementType: AnnouncementType.Profile,
-      fromId: userId,
-      handle: "unknown",
-    };
-
-  const relations =
-    listStatus === ListStatus.FOLLOWERS
-      ? followingDisplayUser
-      : listStatus === ListStatus.FOLLOWING
-      ? followedByDisplayUser
-      : {};
-
-  const connectionsList = Object.keys(relations)
-    .map(profileForId)
-    .filter(
-      (profile) =>
-        relations[profile.fromId]?.status !== RelationshipStatus.UNFOLLOWING
-    );
 
   return (
     <>
